Rename and prepare for major changes
The next few commits will change the input format of the TSVs
and as such should be treated as a 2.0 version.

Namely, there will be a new column added (see changes to blank.tsv)
and some slight renaming will occur.
The next commits will update the code to handle new column names
and the Columbus list to match the new expectations.